fix: Merging lexical and runtime scope in `fn` form so lexical scope actually works
Until now the calling convention would only hand in the current runtime scope for lookups; which works, until you export and call a function from another module where the runtime scope isn't the same as the defining module scope. The current solution to introduce an actual JS closure is really ugly, though. We'll have to think about this at some point but it's enough for our MVP.